requirements.txt
pandas>=2.2
numpy>=1.26
scikit-learn>=1.5
lightgbm>=4.3
xgboost>=2.1
ta>=0.11
scipy>=1.13
joblib>=1.4
pyyaml>=6.0
python-dotenv>=1.0
MetaTrader5>=5.0
loguru>=0.7
optuna>=3.6